Improving build performance and upgrading doxygen
This change started as a simple upgrade to doxygen in the toolshed but build times were so bad I spent more time adding volume caches to the toolshed build and parallelizing the build of doxygen. This also led to some adaptation in procedure as Docker Desktop evolves in its handling of multi-platform images (see updates to READMEs).